+++ src/app/billing/create-quote/page.tsx
@@ -2,7 +2,7 @@ import FillFormContainer from '@/components/ui/containers/fill-form-container';
 import FormContainer from '@/components/ui/containers/form-container';
 import FormHeader from '@/components/ui/header/form-header';
 import { CreateQuoteForm } from '@/components/ui/stripe-forms/stripe-quote-form/create-quote-form';
-import { isOrgAdmin } from "@/lib/server-funtions/clerk";
+import { isOrgAdmin } from "@/lib/server-functions/clerk";
 
 export default async function SendQuotePage() {
   const { orgId, userId } = await isOrgAdmin();
