+++ src/app/billing/manage/quotes/page.tsx
@@ -4,7 +4,7 @@ import FormHeader from "@/components/ui/header/form-header";
 import { ManageQuoteCardView } from "@/components/ui/manage/quotes/manage-quotes-card-view";
 import { PaginationTabs } from "@/components/ui/pagination/pagination-tabs";
 import { fetchQuotes } from "@/lib/dal/stripe-dal";
-import { parseClientListParams } from "@/lib/server-funtions/params";
+import { parseClientListParams } from "@/lib/server-functions/params";
 import { SearchParams } from "@/types/types-params";
 import { Suspense } from "react";
 import SearchFormFallBack from "@/components/ui/fallbacks/search/search-form-fallback";
