+++ src/app/lists/client/page.tsx
@@ -5,9 +5,9 @@ import SearchForm from "@/components/ui/search/search-form";
 import FormContainer from "@/components/ui/containers/form-container";
 import FormHeader from "@/components/ui/header/form-header";
 import { fetchOrgMembers } from "@/lib/dal/dal-org";
-import { isOrgAdmin } from "@/lib/server-funtions/clerk";
+import { isOrgAdmin } from "@/lib/server-functions/clerk";
 import { Suspense } from "react";
-import { parseClientListParams } from "@/lib/server-funtions/params";
+import { parseClientListParams } from "@/lib/server-functions/params";
 import { fetchAllClients } from "@/lib/dal/clients-dal";
 import { SearchParams } from "@/types/types-params";
 import SearchFormFallBack from "@/components/ui/fallbacks/search/search-form-fallback";
