+++ src/components/ui/navigation-menu.tsx
@@ -3,7 +3,7 @@ import * as NavigationMenuPrimitive from "@radix-ui/react-navigation-menu"
 import { cva } from "class-variance-authority"
 import { ChevronDownIcon } from "lucide-react"
 
-import { cn } from "@/lib/server-funtions/utils"
+import { cn } from "@/lib/server-functions/utils"
 
 function NavigationMenu({
   className,
