+++ src/lib/actions/clerk-actions.ts
@@ -1,5 +1,5 @@
 'use server'
-import { isOrgAdmin } from "@/lib/server-funtions/clerk";
+import { isOrgAdmin } from "@/lib/server-functions/clerk";
 
 export async function getIsAdminAction(): Promise<boolean> {
     const { isAdmin } = await isOrgAdmin();
