+++ src/lib/actions/snow-action.ts
@@ -1,6 +1,6 @@
 'use server'
 import { toggleSnowClientDb, assignSnowClearingDb } from "@/lib/DB/db-clients";
-import { isOrgAdmin } from "@/lib/server-funtions/clerk";
+import { isOrgAdmin } from "@/lib/server-functions/clerk";
 import { schemaToggleSnowClient, schemaAssignSnowClearing } from "@/lib/zod/schemas";
 
 export async function toggleSnowClient(clientId: number) {
