+++ src/lib/dal/novu-dal.ts
@@ -5,32 +5,31 @@ const novu = new Novu({
     secretKey: process.env.NOVU_SECRET_KEY,
 });
 
-export async function sayHello(novuId: string, email?: string, userName?: string) {
-    auth.protect()
-    try {
-        const firstName = userName?.split(" ")[0]
-        const lastName = userName?.split(" ")[1]
-        await novu.trigger({
-            workflowId: 'hello-from-landscape-friend',
-            to: {
-                subscriberId: novuId,
-                email,
-                firstName,
-                lastName,
-                timezone: 'America/Edmonton',
-            },
-            payload: {},
-        });
+// export async function sayHello(novuId: string, email?: string, userName?: string) {
+//         try {
+//         const firstName = userName?.split(" ")[0]
+//         const lastName = userName?.split(" ")[1]
+//         await novu.trigger({
+//             workflowId: 'hello-from-landscape-friend',
+//             to: {
+//                 subscriberId: novuId,
+//                 email,
+//                 firstName,
+//                 lastName,
+//                 timezone: 'America/Edmonton',
+//             },
+//             payload: {},
+//         });
 
-    } catch (error) {
-        console.error(error);
-        // return NextResponse.json({ error: 'Failed to trigger Novu workflow' }, { status: 500 });
-    }
-}
+//     } catch (error) {
+//         console.error(error);
+//         // return NextResponse.json({ error: 'Failed to trigger Novu workflow' }, { status: 500 });
+//     }
+// }
 
 
-export async function triggerNotifaction(novuId: string, action: string) {
-    auth.protect()
+export async function triggerNotification(novuId: string, action: string) {
+    await auth.protect()
     try {
         await novu.trigger({
             workflowId: action,
